added new env. variable: ELEV. This sets the flasher LED elevation angle directly (instead of as a correction to the nominal values). This still uses the internal sensor LED coordinates for DOM LEDs, possibly to be changed in the future. So, consider this option experimental for now.
allow definition of the ice layers with just one layer in the icemodel.dat file.
This selects the ice layer thickness of 100 m (since this cannot be calculated
from the depth of just a single layer).

llh/DirectFit:

Implemented a fix to ERT last bin problem (where only half of simulation was
filled into the last time bin preceeding the ERT exclusion window). This is needed
for the pre-binned data (such as flasher data used for ice fitting).

Set the new env. variable MONO=1 to bin data such that the bins rise and fall
monotonically around a single maximum bin. This is done via a modification to the
bayesian block algorithm. This works well for flasher data, but you may need to
disable this (default behavior) if multiple peaks in data waveforms are important
(and the fit needs to reproduce this - such as in tau double bang studies).

Set the new env. variable PEAK=1 to use a slower bin optimization that searches
for the best peak bin location in bayesian-block single-peak bin solutions.
Otherwise (nominally) the peak bin is assumed to contain the max value bin of the
input binning. The captured binned range extends (up to) 500 ns from the front of
the peak bin into earlier times, and (up to) 1000 ns from the end of the peak bin
into later times. This can be used with or without MONO=1 option.

Changed the behavior of including the overflow data into the earliest and latest
bins. This is no longer done, the data that is outside the binned range is now
ignored. This is controlled by the boolean variable keep=false (if you'd like to
change it back).