llh/DirectFit:
fixed noise term for both keep=true/false. For the former the overflow bin noise
was often not extended to the overflow bins for the first bin (last bin was ok).
The latter was not adjusted correctly to omit noise from overflow bins.

Waveform printout has been amended to print unused later part of the waveforms.
This is needed to calculate absolute bin times for later plotting (w/keep=false)